requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.1
python-louvain>=0.16
requests>=2.31.0
python-dotenv>=1.0.0
wordcloud>=1.9.0
Pillow>=10.0.0
scikit-learn>=1.3.0
rich>=13.0.0